Move `dot_ix::model` into `dot_ix_model` crate.
