+++ crate/model/src/common.rs
