+++ crate/model/src/common/dot_src_and_styles.rs
