+++ crate/model/src/common/edge.rs
