+++ crate/model/src/common/graphviz_dot_theme.rs
