+++ crate/model/src/common/id_newtype.rs
