+++ crate/model/src/info_graph/node_info.rs
