+++ src/lib.rs
@@ -1,7 +1,6 @@
 use cfg_if::cfg_if;
 pub mod app;
 pub mod fileserv;
-pub mod model;
 pub mod rt;
 
 cfg_if! { if #[cfg(feature = "hydrate")] {
